fix: replace remaining deprecated calls and modulo float operands for PHP 8.5
Adds .phan/config.php (target PHP 8.5, analyzes library/) to run Phan
alongside PHPStan's deprecation-rules checks.

Fixes flagged by Phan:
- Zend_Reflection_Parameter::getClass() called the deprecated parent
  ReflectionParameter::getClass(); reimplemented via getType().
- Zend_Ldap::connect() called ldap_connect($host, $port), deprecated as of
  PHP 8.3; always connects via the already-built URI string instead.
- Unparenthesized nested ternary in Zend_Service_Console_Command, deprecated
  since PHP 8.0.
- Zend_Soap_Client::_doRequest() invoked SoapClient::__doRequest with an
  alternate class scope via a "Class::method" callable, deprecated in
  PHP 8.2; now uses ReflectionMethod::invoke().
- ~30 uses of the modulo operator on float operands (mostly ceil()/floor()
  results, or ints inferred as float by string-to-number coercion), deprecated
  since PHP 8.1, across Barcode, Pdf, Text, Locale and Validate. Cast the
  operand to int at each site.

+++ library/Zend/Pdf/Cmap/SegmentToDelta.php
@@ -170,7 +170,7 @@ class Zend_Pdf_Cmap_SegmentToDelta extends Zend_Pdf_Cmap
                 /* This segment uses a simple mapping from character code to
                  * glyph number.
                  */
-                $glyphNumbers[$key] = ($characterCode + $this->_segmentTableIdDeltas[$subtableIndex]) % 65536;
+                $glyphNumbers[$key] = ((int) ($characterCode + $this->_segmentTableIdDeltas[$subtableIndex])) % 65536;
 
             } else {
                 /* This segment relies on the glyph index array to determine the
@@ -234,7 +234,7 @@ class Zend_Pdf_Cmap_SegmentToDelta extends Zend_Pdf_Cmap
         }
 
         if ($this->_segmentTableIdRangeOffsets[$subtableIndex] == 0) {
-            $glyphNumber = ($characterCode + $this->_segmentTableIdDeltas[$subtableIndex]) % 65536;
+            $glyphNumber = ((int) ($characterCode + $this->_segmentTableIdDeltas[$subtableIndex])) % 65536;
         } else {
             $glyphIndex = ($characterCode - $this->_segmentTableStartCodes[$subtableIndex] +
                            $this->_segmentTableIdRangeOffsets[$subtableIndex] - $this->_segmentCount +
@@ -284,7 +284,7 @@ class Zend_Pdf_Cmap_SegmentToDelta extends Zend_Pdf_Cmap
                 for ($code =  $this->_segmentTableStartCodes[$segmentNum];
                      $code <= $this->_segmentTableEndCodes[$segmentNum];
                      $code++) {
-                    $glyphNumbers[$code] = ($code + $delta) % 65536;
+                    $glyphNumbers[$code] = ((int) ($code + $delta)) % 65536;
                 }
             } else {
                 $code       = $this->_segmentTableStartCodes[$segmentNum];
